Fixed compiler warnings.

+++ src/utils/vec3.cpp
@@ -41,8 +41,8 @@ void Vec3::setHPR(const btQuaternion& q)
  */
 void Vec3::setPitchRoll(const Vec3 &normal)
 {
-    const float X = sin(getHeading());
-    const float Z = cos(getHeading());
+    const float X = sinf(getHeading());
+    const float Z = cosf(getHeading());
     // Compute the angle between the normal of the plane and the line to
     // (x,0,z).  (x,0,z) is normalised, so are the coordinates of the plane,
     // which simplifies the computation of the scalar product.
